requirements.txt
django==4.2.11
django-environ==0.11.2
django-cors-headers==4.3.1
djangorestframework==3.15.1
pillow==10.3.0
django-storages==1.14.2
django-ckeditor==6.7.1
psycopg2==2.9.9

argon2-cffi==23.1.0
djoser==2.2.3
djangorestframework-simplejwt==5.3.1
social-auth-app-django==5.4.1
PyJWT==2.8.0